+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.93.0
 - Support tracing of deferred HTTP exits (that is, capture HTTP exits if the entry that triggered it has already finished).
 - [AWS Lambda] Add support for INSTANA_DISABLE_CA_CHECK.